requirements.txt
langchain>=0.2
langchain-google-genai>=1.0
langchain-community>=0.2
langgraph>=0.1
chromadb>=0.5
tavily-python>=0.3
pypdf>=4.0
python-dotenv>=1.0
pydantic>=2.0
sentence-transformers>=3.0
